Move API endpoints and schema objects outside of server.py (#65)
* Add input/output schema in action_run, action_test, trigger_test method's docstrings

* Update doc strings to use body and specify array type

* Revert the changes for ActionTriggerOutputSchema response

* Add missed parameter to run_action_trigger method

* Add versioning using blueprint

* Update unittest test_server_spec

* Modifications to support legacy and versioned endpoints (#63)

* Modifications to support legacy and versioned endpoints without code duplication; normalizing blueprint names

* Remove registration of legacy methods to APISpec


* Move API endpoints and schema objects outside of server.py

* Merge conflicts

* Removed extra empty line from endpoints.py

* Code review changes

* Break long line

* Remove redundant code
